Use same tag treatment function
